add delayed_paperclip and carousel

+++ app/controllers/rooms_controller.rb
@@ -1,4 +1,5 @@
 class RoomsController < ApplicationController
+  
   def index
     @rooms = Room.all
   end
@@ -9,11 +10,13 @@ class RoomsController < ApplicationController
   end
 
   def new
+    check_role
     @room = Room.new
     @room.images.build
   end
 
   def create
+    check_role
     @room = current_user.rooms.build(room_params)
     if @room.save
       redirect_to @room, notice: 'Room saved'
@@ -25,11 +28,12 @@ class RoomsController < ApplicationController
 
   def edit
     @room = Room.find(params[:id])
-
+    check_owner(@room)
   end
 
   def update
     @room = Room.find(params[:id])
+    check_owner(@room)
     if @room.update(room_params)
       flash[:notice] = "The room was updated"
       redirect_to @room, notice: 'Room saved'
@@ -41,6 +45,7 @@ class RoomsController < ApplicationController
 
   def destroy
     @room = Room.find(params[:id])
+    check_owner(@room)
     if @room.destroy
       flash[:notice] = "The room was deleted"
       redirect_to root_path
@@ -54,5 +59,18 @@ class RoomsController < ApplicationController
   def room_params
     params.require(:room).permit(:title, :description, :city, :country, :latitude, :longitude,
       images_attributes: [:file])
+  end
+  
+  def check_role
+    if current_user.role != 'host'
+      raise 'You are not a host'
+    end
+  end
+  
+  def check_owner(room)
+    if room.user_id != current_user.id
+      raise 'You are not the owner'
+      render root_path
+    end
   end
 end
